requirements.txt
argparse
numpy
torch
torchvision
numpy-quaternion
tqdm
Pillow
pickle-mixin
mat73